+++ src/main/java/forestry/apiculture/items/ItemBlockHoneyComb.java
@@ -3,9 +3,6 @@ package forestry.apiculture.items;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-
 import forestry.api.core.ItemGroups;
 import forestry.apiculture.blocks.BlockHoneyComb;
 import forestry.core.items.ItemBlockForestry;
@@ -13,10 +10,9 @@ import forestry.core.items.definitions.IColoredItem;
 
 public class ItemBlockHoneyComb extends ItemBlockForestry<BlockHoneyComb> implements IColoredItem {
 	public ItemBlockHoneyComb(BlockHoneyComb block) {
-		super(block, new Item.Properties().tab(ItemGroups.tabApiculture));
+		super(block, new Item.Properties().tab(block.type.isUnused() ? null : ItemGroups.tabApiculture));
 	}
 
-	@OnlyIn(Dist.CLIENT)
 	@Override
 	public int getColorFromItemStack(ItemStack stack, int tintIndex) {
 		EnumHoneyComb honeyComb = getBlock().getType();